# stays-platform/spec/swagger.json
{
	"components": {
...
		"schemas": {
			"Entity": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "number",
						"format": "double"
					},
					"updatedAt": {
						"type": "number",
						"format": "double"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Address": {
				"properties": {
					"city": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"address1": {
						"type": "string"
					},
					"address2": {
						"type": "string"
					},
					"zip": {
						"type": "number",
						"format": "double"
					},
					"country": {
						"type": "string"
					}
				},
				"required": [
					"city",
					"state",
					"address1",
					"address2",
					"zip",
					"country"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Coordinates": {
				"properties": {
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"latitude",
					"longitude"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Location": {
				"properties": {
					"address": {
						"$ref": "#/components/schemas/Address"
					},
					"coordinates": {
						"$ref": "#/components/schemas/Coordinates"
					},
					"region": {
						"type": "string"
					}
				},
				"required": [
					"address",
					"coordinates",
					"region"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Photo": {
				"properties": {
					"url": {
						"type": "string"
					},
					"priority": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"url",
					"priority",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StayApplicationStatus": {
				"enum": [
					"Pending",
					"Rejected",
					"Accepted"
				],
				"type": "string"
			},
			"Social": {
				"properties": {
					"partner": {
						"type": "string"
					},
					"link": {
						"type": "string"
					}
				},
				"required": [
					"partner",
					"link"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Booking": {
				"properties": {
					"partner": {
						"type": "string"
					},
					"link": {
						"type": "string"
					}
				},
				"required": [
					"partner",
					"link"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EarlyBooking": {
				"properties": {
					"startDate": {
						"type": "number",
						"format": "double"
					},
					"endDate": {
						"type": "number",
						"format": "double"
					},
					"bookingLink": {
						"type": "string"
					}
				},
				"required": [
					"startDate",
					"endDate",
					"bookingLink"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Stay": {
				"properties": {
					"name": {
						"type": "string"
					},
					"enable": {
						"type": "boolean"
					},
					"description": {
						"type": "string"
					},
					"location": {
						"$ref": "#/components/schemas/Location"
					},
					"currentRate": {
						"type": "number",
						"format": "double"
					},
					"averageRate": {
						"type": "number",
						"format": "double"
					},
					"capacity": {
						"type": "number",
						"format": "double"
					},
					"bedrooms": {
						"type": "number",
						"format": "double"
					},
					"petsAllowed": {
						"type": "boolean"
					},
					"onSiteParking": {
						"type": "boolean"
					},
					"type": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"specialInterests": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"amenities": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"photos": {
						"items": {
							"$ref": "#/components/schemas/Photo"
						},
						"type": "array"
					},
					"status": {
						"$ref": "#/components/schemas/StayApplicationStatus"
					},
					"demand": {
						"type": "string"
					},
					"social": {
						"items": {
							"$ref": "#/components/schemas/Social"
						},
						"type": "array"
					},
					"booking": {
						"items": {
							"$ref": "#/components/schemas/Booking"
						},
						"type": "array"
					},
					"favoriteCount": {
						"type": "number",
						"format": "double"
					},
					"organizationId": {
						"type": "string"
					},
					"earlyBooking": {
						"$ref": "#/components/schemas/EarlyBooking"
					}
				},
				"required": [
					"name",
					"enable",
					"description",
					"location",
					"currentRate",
					"averageRate",
					"capacity",
					"bedrooms",
					"petsAllowed",
					"onSiteParking",
					"type",
					"specialInterests",
					"amenities",
					"tags",
					"photos",
					"status",
					"demand",
					"social",
					"booking",
					"favoriteCount",
					"organizationId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StayRecord": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Stay"
					},
					{
						"$ref": "#/components/schemas/Entity"
					}
				]
			},
...
			"Org": {
				"properties": {
					"name": {
						"type": "string"
					},
					"userIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"stayIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"userIds",
					"stayIds"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OrgRecord": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Entity"
					},
					{
						"$ref": "#/components/schemas/Org"
					}
				]
			},
...
			"UserMembership": {
				"enum": [
					"Standard",
					"Premium"
				],
				"type": "string"
			},
			"Role": {
				"enum": [
					"Stayer",
					"Host",
					"Employee",
					"Admin"
				],
				"type": "string"
			},
			"User": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"enabled": {
						"type": "boolean"
					},
					"email": {
						"type": "string"
					},
					"userMembership": {
						"$ref": "#/components/schemas/UserMembership"
					},
					"lastActive": {
						"type": "number",
						"format": "double"
					},
					"roles": {
						"items": {
							"$ref": "#/components/schemas/Role"
						},
						"type": "array"
					},
					"favorites": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"enabled",
					"email",
					"userMembership",
					"lastActive",
					"roles",
					"favorites"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserRecord": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Entity"
					},
					{
						"$ref": "#/components/schemas/User"
					}
				]
			},
...
		},
		"securitySchemes": {
			"user": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"info": {
		"title": "stays-platform",
		"version": "1.0.0",
		"description": "back end for stays app",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Nick Schrock"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/orgs": {
			"get": {
				"operationId": "GetOrgs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/OrgRecord"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"user": [
							"Admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "PostOrg",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OrgRecord"
								}
							}
						}
					}
				},
				"security": [
					{
						"user": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Org"
							}
						}
					}
				}
			}
		},
		"/orgs/{orgId}": {
			"get": {
				"operationId": "GetOrg",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OrgRecord"
								}
							}
						}
					}
				},
				"security": [
					{
						"user": [
							"Host"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "orgId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "EditOrg",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"user": [
							"Host"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "orgId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Org"
							}
						}
					}
				}
			}
		},
		"/orgs/{orgId}/users/{userId}": {
			"put": {
				"operationId": "AddUserToOrg",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"user": [
							"Host"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "orgId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "RemoveUserFromOrg",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"user": [
							"Host"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "orgId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/orgs/{orgId}/stays": {
			"get": {
				"operationId": "GetOrgsStays",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/StayRecord"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"user": [
							"Host"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "orgId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/orgs/{orgId}/stays/{stayId}": {
			"put": {
				"operationId": "AddStayToOrg",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"user": [
							"Host"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "orgId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "stayId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "RemoveStayFromOrg",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"user": [
							"Host"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "orgId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "stayId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/users/self": {
			"get": {
				"operationId": "GetSelf",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserRecord"
								}
							}
						}
					}
				},
				"security": [
					{
						"user": [
							"Stayer"
						]
					}
				],
				"parameters": []
			}
		},
		"/users/self/orgs": {
			"get": {
				"operationId": "GetUserOrgs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/OrgRecord"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"user": [
							"Host"
						]
					}
				],
				"parameters": []
			}
		},
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserRecord"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"user": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserRecord"
								}
							}
						}
					}
				},
				"security": [
					{
						"firebase": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					}
				}
			}
		},
		"/users/{userId}": {
			"post": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserRecord"
								}
							}
						}
					}
				},
				"security": [
					{
						"user": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "PatchUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"user": [
							"Admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/self/favorites/{stayId}": {
			"post": {
				"operationId": "AddFavorite",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"user": [
							"Stayer"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "stayId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "RemoveFavorite",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"user": [
							"Stayer"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "stayId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}